impl review suggestions
