+++ raphtory-api/Cargo.toml
@@ -21,7 +21,14 @@ dashmap = { workspace = true }
 rustc-hash = { workspace = true }
 lock_api = { workspace = true }
 parking_lot = { workspace = true }
-pyo3 = { workspace = true }
+pyo3 = { workspace = true, optional = true }
 rayon = { workspace = true }
 rand = { workspace = true }
 quickcheck_macros = { workspace = true }
+
+[features]
+default = []
+# Enables generating the pyo3 python bindings
+python = [
+    "dep:pyo3",
+]
