+++ raphtory-api/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod core;
+
+#[cfg(feature = "python")]
+pub mod python;
