+++ raphtory-api/src/python/mod.rs
@@ -1,0 +1,20 @@
+use pyo3::*;
+use crate::core::storage::arc_str::ArcStr;
+
+impl IntoPy<PyObject> for ArcStr {
+    fn into_py(self, py: Python<'_>) -> PyObject {
+        self.0.into_py(py)
+    }
+}
+
+impl ToPyObject for ArcStr {
+    fn to_object(&self, py: Python) -> PyObject {
+        self.0.to_string().to_object(py)
+    }
+}
+
+impl<'source> FromPyObject<'source> for ArcStr {
+    fn extract(ob: &'source PyAny) -> PyResult<Self> {
+        ob.extract::<String>().map(|v| v.into())
+    }
+}
